module-4-dockerized-mlflow-gcp/requirements.txt
mlflow==1.28.0
python-dotenv==0.21.0
